New options: custom message, and put enrolled users in a group #14
This change adds two new option for this plugin:

First, there is the ability to have a custom message, which is displayed
instead of the standard 'This course requires a payment for entry.'
text. This is stored in customtext1. If it is blank, then the standard
text is still used.

Second, there is an option to select a group. If this is selected, then
when a user is enrolled in the course, they are put into that group.

Finally, I made some changes to the HTML that is displayed on the Enrol
options page when a user tries to enter a course.
 * I swapped the order of the instancename, and the message. The
   instancename looks like a heading, that seemed more natural. Note
   that Moodle's standard self-enrolment plugin now puts the instance
   name before the message.
 * I change the instance name to acutally be a heading. This is better
   for accessibility. I also think it looks better.
 * I changed the hard-coded align="center" to align="center".

I also  changed the size of some of the input boxes on the edit form
(Custom instance name and Max Enrolled users). I hope the new sizes are
better.